fix branch extend commit_id
